Move get_data from tests to cwltool
