+++ tests/test_examples.py
@@ -8,7 +8,7 @@ import sys
 from io import StringIO
 
 from cwltool.errors import WorkflowException
-from cwltool.utils import onWindows
+from cwltool.utils import onWindows, get_data
 
 try:
     reload
@@ -26,7 +26,6 @@ import cwltool.workflow
 import schema_salad.validate
 from cwltool.main import main
 
-from .util import get_data
 
 sys.argv = ['']
