+++ tests/test_ext.py
@@ -10,8 +10,7 @@ import cwltool.pathmapper
 import cwltool.process
 import cwltool.workflow
 from cwltool.main import main
-from cwltool.utils import onWindows
-from .util import get_data
+from cwltool.utils import onWindows, get_data
 
 
 @pytest.mark.skipif(onWindows(),
