+++ tests/test_pack.py
@@ -17,8 +17,7 @@ from cwltool import load_tool
 from cwltool.load_tool import fetch_document, validate_document
 from cwltool.main import makeRelative, main, print_pack
 from cwltool.pathmapper import adjustDirObjs, adjustFileObjs
-from cwltool.utils import onWindows
-from .util import get_data
+from cwltool.utils import onWindows, get_data
 
 
 class TestPack(unittest.TestCase):
