+++ tests/test_toolargparse.py
@@ -4,9 +4,7 @@ import pytest
 from tempfile import NamedTemporaryFile
 
 from cwltool.main import main
-from cwltool.utils import onWindows
-
-from .util import get_data
+from cwltool.utils import onWindows, get_data
 
 class ToolArgparse(unittest.TestCase):
     script = '''
